ingest/requirements.txt
h5py
numpy
umap-learn
requests